Backend/requirements.txt
# FastAPI framework
fastapi

# ASGI server
uvicorn[standard]

# Async PostgreSQL driver
asyncpg

# SQLAlchemy for async ORM
SQLAlchemy

# Pydantic v2 (with email validator)
pydantic[email]
pydantic-settings

# Password hashing
passlib[bcrypt]

# JWT handling
python-jose[cryptography]

# Alembic for migrations
alembic

# Environment variable loader
python-dotenv

# Testing
pytest
pytest-asyncio
httpx